+++ SandboxiePlus/SandMan/Helpers/WinHelper.h
@@ -1,0 +1,5 @@
+#pragma once
+
+QImage LoadWindowsIcon(const QString& Path, quint32 Index);
+
+bool PickWindowsIcon(QWidget* pParent, QString& Path, quint32& Index);